Fix requests when changing location
I didn't pay attention to the page state changing when the component was re-rendered.
When a component is removed from the DOM, the page state reverts to its initial value which spoils next request during dynamic pagination.
Conclusion: we need to create the page state in the reducer, also to create the action creator for this.